:sparkles: Finishes pagination feature

+++ StorageApi/StorageApi/Controllers/Models/Response/PaginatedResponse.cs
@@ -4,6 +4,6 @@
     {
         public int PageNumber { get; set; } = 1;
         public int PageSize { get; set; } = 10;
-        public int ItemCount { get; set; }
+        public long ItemCount { get; set; }
     }
 }
